moved slide-toggle styles to overrides, removed switch component
