+++ libs/design-system/styles/_theming.scss
@@ -16,4 +16,5 @@
   @include overrides.menu-overrides();
   @include overrides.slider-overrides();
   @include overrides.select-overrides();
+  @include overrides.slide-toggle-overrides();
 }
