+++ libs/design-system/styles/overrides/_index.scss
@@ -7,6 +7,7 @@
 @forward './radio-button' as radio-button-*;
 @forward './select' as select-*;
 @forward './slider' as slider-*;
+@forward './slide-toggle' as slide-toggle-*;
 
 @mixin overrides() {
   @include button-overrides();
@@ -18,4 +19,5 @@
   @include radio-button-overrides();
   @include select-overrides();
   @include slider-overrides();
+  @include slide-toggle-overrides();
 }
