--- libs/design-system/switch-button/ng-package.json
@@ -1,5 +1,0 @@
-{
-  "lib": {
-    "entryFile": "src/index.ts"
-  }
-}
