--- libs/design-system/switch-button/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/switch.component';
