--- libs/design-system/switch-button/src/lib/switch.component.html
@@ -1,7 +1,0 @@
-<mat-slide-toggle
-  [checked]="isChecked()"
-  (change)="onSwitchToggle($event)"
-  [disabled]="isDisabled()"
-  [hideIcon]="hideIcon()"
->
-</mat-slide-toggle>
